chore: refactor component
